requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
openai==1.54.0
python-docx==1.1.2
openpyxl==3.1.5
pypdf==6.0.0
pydantic==2.9.2
pydantic-settings==2.6.0
pywebview==5.3.1
requests==2.32.3
trafilatura==2.0.0
tiktoken>=0.12.0
slowapi==0.1.9
